+++ CHANGELOG.md
@@ -3,6 +3,7 @@ v2.1.2
 
 * Fixed regression where `cli` was not accepting hyphens for site names. [#1711](https://github.com/kalabox/kalabox/issues/1711)
 * Updated to latest stable Docker for Mac and Docker for Windows `v1.12.3`. [#1711](https://github.com/kalabox/kalabox/issues/1711)
+* Fixed bug where non-Kalabox containers were getting wiped out during Kalabox cleanup. [#1708](https://github.com/kalabox/kalabox/issues/1708)
 
 v2.1.1
 ======